BugID:21767265: Change vfs_lseek the parameter type

+++ kernel/fs/vfs/vfs.c
@@ -267,7 +267,7 @@ int32_t vfs_do_pollfd(int32_t fd, int32_t flag, vfs_poll_notify_t notify,
     return ret;
 }
 
-uint32_t vfs_lseek(int32_t fd, uint32_t offset, int32_t whence)
+uint32_t vfs_lseek(int32_t fd, int64_t offset, int32_t whence)
 {
     int32_t ret = VFS_ERR_NOSYS;
